feat: centralize rng logic for insight browser
