+++ alpha_factory_v1/demos/alpha_agi_insight_v1/insight_browser_v1/build.js
@@ -32,6 +32,10 @@ async function bundle() {
   await fs.copyFile('d3.v7.min.js', `${OUT_DIR}/d3.v7.min.js`);
   await fs.copyFile('lib/bundle.esm.min.js', `${OUT_DIR}/bundle.esm.min.js`);
   await fs.copyFile('lib/pyodide.js', `${OUT_DIR}/pyodide.js`);
+  await fs.mkdir(`${OUT_DIR}/worker`, { recursive: true });
+  await fs.copyFile('worker/evolver.js', `${OUT_DIR}/worker/evolver.js`);
+  await fs.mkdir(`${OUT_DIR}/src/utils`, { recursive: true });
+  await fs.copyFile('src/utils/rng.js', `${OUT_DIR}/src/utils/rng.js`);
 
   const sha384 = async (file) => {
     const data = await fs.readFile(`${OUT_DIR}/${file}`);
