+++ alpha_factory_v1/demos/alpha_agi_insight_v1/insight_browser_v1/index.html
@@ -69,16 +69,7 @@ import {initFpsMeter} from './src/ui/fpsMeter.js';
 import {initI18n,t} from './src/ui/i18n.js';
 import {chat as llmChat} from './src/utils/llm.js';
 import { initTelemetry } from '../../../../src/telemetry.js';
-
-function lcg(seed){
-  function rand(){
-    seed=Math.imul(1664525,seed)+1013904223>>>0;
-    return seed/2**32;
-  }
-  rand.state=()=>seed;
-  rand.set=s=>{seed=s>>>0;};
-  return rand;
-}
+import { lcg } from './src/utils/rng.js';
 
 let panel,pauseBtn,exportBtn,dropZone
 let criticPanel,logicCritic,feasCritic
