+++ alpha_factory_v1/demos/alpha_agi_insight_v1/insight_browser_v1/worker/evolver.js
@@ -1,16 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import { mutate } from '../src/evolve/mutate.js';
 import { paretoFront } from '../src/utils/pareto.js';
-
-function lcg(seed) {
-  function rand() {
-    seed = Math.imul(1664525, seed) + 1013904223 >>> 0;
-    return seed / 2 ** 32;
-  }
-  rand.state = () => seed;
-  rand.set = (s) => { seed = s >>> 0; };
-  return rand;
-}
+import { lcg } from '../src/utils/rng.js';
 
 function shuffle(arr, rand) {
   for (let i = arr.length - 1; i > 0; i--) {
